+++ Editor/Attributes/CustomContextMenuAttribute.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace NewGraph {
+    [AttributeUsage(AttributeTargets.Class,  AllowMultiple = false)]
+    public class CustomContextMenuAttribute : Attribute {
+
+        public CustomContextMenuAttribute() { }
+    }
+}