+++ Editor/Views/GraphView.cs
@@ -27,6 +27,7 @@ namespace NewGraph {
             this.OnAction = OnAction;
         }
 
+
         private void OnKeyDown(Event evt) {
             ExecuteShortcutHandler(evt.keyCode, evt.modifiers);
         }